// src/components/Publication/types.ts
export enum Theme {
  default = "default",
  light = "light",
  dark = "dark",
  mint = "mint",
  green = "green",
  peach = "peach",
  lavender = "lavender",
  blonde = "blonde",
}

export enum ThemeColor {
  lightGray = "#dcdcdc",
  mediumGray = "#969696",
  darkGray = "#464646",
  lightBlack = "#1C1D1C",
  lightGreen = "#75B680",
  mint = "#bce0c7",
  green = "#364239",
  peach = "#f5d4d2",
  lavender = "#dbccf3",
  blonde = "#ffebB8",
  white = "white",
  black = "black",
  transparent = "none",
  red = "red",
}